+++ config.h
@@ -88,6 +88,7 @@ typedef enum {
 typedef struct struct_config {
   int message_interval;
   int send_usb_bulk;
+  bool send_midi_monitor;
   int send_motion_interval;
   int send_motion_14bit;
   int send_button_14bit;
@@ -116,6 +117,7 @@ extern config_t config;
 config_t config = {
   .message_interval = 1,      // interval in ms
   .send_usb_bulk = 0,         // send Striso binary protocol
+  .send_midi_monitor = 0,     // monitor MIDI in over Striso protocol
   .send_motion_interval = 127,// 0 = disable, 127 only internal, else x10ms
   .send_motion_14bit = 0,     // send 14 bit motion CC
   .send_button_14bit = 0,     // send 14 bit MPE CC